assistant/requirements.txt
fastapi
uvicorn
pydantic
requests
sentence-transformers
chromadb
python-docx
